requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dj-rest-auth==6.0.0
Django==4.2
django-allauth==64.1.0
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
idna==3.8
packaging==24.1
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
secure==0.3.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
whitenoise==6.7.0
